fix(icons): rename bubble as chat and grouping
