+++ packages/aksara-icons/src/chat/IconFAQ.tsx
