+++ packages/aksara-icons/src/chat/IconPhrase.tsx
