+++ packages/aksara-icons/src/chat/IconTranscript.tsx
