+++ packages/aksara-icons/src/chat/IconWaba.tsx
