+++ packages/aksara-icons/src/chat/bot/IconMessageBot.tsx
