+++ packages/aksara-icons/src/chat/bot/index.ts
