+++ packages/aksara-icons/src/chat/bubble/index.ts
@@ -1,5 +1,3 @@
-export * from './conversation';
-export * from './tooltip';
 export { default as IconBubble } from './IconBubble';
 export { default as IconBubble2 } from './IconBubble2';
 export { default as IconBubble3 } from './IconBubble3';
@@ -13,13 +11,3 @@ export { default as IconBubbleEdit } from './IconBubbleEdit';
 export { default as IconBubbleLove } from './IconBubbleLove';
 export { default as IconBubblePlus } from './IconBubblePlus';
 export { default as IconBubbleTick } from './IconBubbleTick';
-export { default as IconEntities } from './IconEntities';
-export { default as IconFAQ } from './IconFAQ';
-export { default as IconNLUs } from './IconNLUs';
-export { default as IconPhrase } from './IconPhrase';
-export { default as IconTranscript } from './IconTranscript';
-export { default as IconUnchat } from './IconUnchat';
-export { default as IconWaba } from './IconWaba';
-export { default as IconMessageBot } from './bot';
-export { default as IconMessageTemplate } from './message';
-export { default as IconNLUTrait } from './nlu';
