+++ packages/aksara-icons/src/chat/conversation/IconConversation2.tsx
