+++ packages/aksara-icons/src/chat/conversation/IconConversation3.tsx
