+++ packages/aksara-icons/src/chat/index.ts
@@ -1,0 +1,14 @@
+export * from './bubble';
+export * from './conversation';
+export * from './tooltip';
+
+export { default as IconEntities } from './IconEntities';
+export { default as IconFAQ } from './IconFAQ';
+export { default as IconNLUs } from './IconNLUs';
+export { default as IconPhrase } from './IconPhrase';
+export { default as IconTranscript } from './IconTranscript';
+export { default as IconUnchat } from './IconUnchat';
+export { default as IconWaba } from './IconWaba';
+export { default as IconMessageBot } from './bot';
+export { default as IconMessageTemplate } from './message';
+export { default as IconNLUTrait } from './nlu';
