+++ packages/aksara-icons/src/chat/message/IconMessageTemplate.tsx
