+++ packages/aksara-icons/src/chat/message/index.ts
