+++ packages/aksara-icons/src/chat/nlu/IconNLUTrait.tsx
