+++ packages/aksara-icons/src/chat/nlu/index.ts
