+++ packages/aksara-icons/src/chat/tooltip/IconTooltipBottom.tsx
