+++ packages/aksara-icons/src/chat/tooltip/IconTooltipLeft.tsx
