+++ packages/aksara-icons/src/chat/tooltip/IconTooltipLove.tsx
