+++ packages/aksara-icons/src/chat/tooltip/IconTooltipRight.tsx
