+++ packages/aksara-icons/src/chat/tooltip/IconTooltipTop.tsx
