+++ packages/aksara-icons/src/chat/tooltip/index.ts
