+++ packages/aksara-icons/src/index.ts
@@ -1,7 +1,7 @@
 export * from './ai';
 export * from './atoms';
 export * from './basic';
-export * from './bubble';
+export * from './chat';
 export * from './chart';
 export * from './doc';
 export * from './media';
